Refs #21082. Move non-exported methods to `BaseWriter`.
