+++ src/cpp/rtps/builtin/liveliness/WLP.cpp
@@ -39,6 +39,7 @@
 #include <rtps/reader/StatefulReader.hpp>
 #include <rtps/resources/ResourceEvent.h>
 #include <rtps/resources/TimedEvent.h>
+#include <rtps/writer/BaseWriter.hpp>
 #include <rtps/writer/LivelinessManager.hpp>
 #include <rtps/writer/StatefulWriter.hpp>
 #include <utils/TimeConversion.hpp>
@@ -620,7 +621,10 @@ bool WLP::add_local_writer(
         const fastdds::dds::WriterQos& wqos)
 {
     std::lock_guard<std::recursive_mutex> guard(*mp_builtinProtocols->mp_PDP->getMutex());
+
     EPROSIMA_LOG_INFO(RTPS_LIVELINESS, W->getGuid().entityId << " to Liveliness Protocol");
+
+    BaseWriter* base_writer = BaseWriter::downcast(W);
 
     double wAnnouncementPeriodMilliSec(fastdds::rtps::TimeConv::Duration_t2MilliSecondsDouble(wqos.m_liveliness.
                     announcement_period));
@@ -650,7 +654,7 @@ bool WLP::add_local_writer(
             }
             automatic_liveliness_assertion_->restart_timer();
         }
-        automatic_writers_.push_back(W);
+        automatic_writers_.push_back(base_writer);
     }
     else if (wqos.m_liveliness.kind == dds::MANUAL_BY_PARTICIPANT_LIVELINESS_QOS)
     {
@@ -677,7 +681,7 @@ bool WLP::add_local_writer(
             }
             manual_liveliness_assertion_->restart_timer();
         }
-        manual_by_participant_writers_.push_back(W);
+        manual_by_participant_writers_.push_back(base_writer);
 
         if (!pub_liveliness_manager_->add_writer(
                     W->getGuid(),
@@ -689,7 +693,7 @@ bool WLP::add_local_writer(
     }
     else if (wqos.m_liveliness.kind == dds::MANUAL_BY_TOPIC_LIVELINESS_QOS)
     {
-        manual_by_topic_writers_.push_back(W);
+        manual_by_topic_writers_.push_back(base_writer);
 
         if (!pub_liveliness_manager_->add_writer(
                     W->getGuid(),
@@ -703,8 +707,6 @@ bool WLP::add_local_writer(
     return true;
 }
 
-typedef std::vector<RTPSWriter*>::iterator t_WIT;
-
 bool WLP::remove_local_writer(
         RTPSWriter* W)
 {
@@ -713,17 +715,18 @@ bool WLP::remove_local_writer(
     EPROSIMA_LOG_INFO(RTPS_LIVELINESS, W->getGuid().entityId << " from Liveliness Protocol");
 
     LivelinessData::WriterStatus writer_status;
+    BaseWriter* base_writer = BaseWriter::downcast(W);
 
-    if (W->get_liveliness_kind() == dds::AUTOMATIC_LIVELINESS_QOS)
+    if (base_writer->get_liveliness_kind() == dds::AUTOMATIC_LIVELINESS_QOS)
     {
         auto it = std::find(
             automatic_writers_.begin(),
             automatic_writers_.end(),
-            W);
+            base_writer);
 
         if (it == automatic_writers_.end())
         {
-            EPROSIMA_LOG_WARNING(RTPS_LIVELINESS, "Writer " << W->getGuid() << " not found.");
+            EPROSIMA_LOG_WARNING(RTPS_LIVELINESS, "Writer " << base_writer->getGuid() << " not found.");
             return false;
         }
 
@@ -749,29 +752,29 @@ bool WLP::remove_local_writer(
         automatic_liveliness_assertion_->update_interval_millisec(min_automatic_ms_);
         return true;
     }
-    else if (W->get_liveliness_kind() == dds::MANUAL_BY_PARTICIPANT_LIVELINESS_QOS)
+    else if (base_writer->get_liveliness_kind() == dds::MANUAL_BY_PARTICIPANT_LIVELINESS_QOS)
     {
         auto it = std::find(
             manual_by_participant_writers_.begin(),
             manual_by_participant_writers_.end(),
-            W);
+            base_writer);
 
         if (it == manual_by_participant_writers_.end())
         {
-            EPROSIMA_LOG_WARNING(RTPS_LIVELINESS, "Writer " << W->getGuid() << " not found.");
+            EPROSIMA_LOG_WARNING(RTPS_LIVELINESS, "Writer " << base_writer->getGuid() << " not found.");
             return false;
         }
 
         manual_by_participant_writers_.erase(it);
 
         if (!pub_liveliness_manager_->remove_writer(
-                    W->getGuid(),
-                    W->get_liveliness_kind(),
-                    W->get_liveliness_lease_duration(),
+                    base_writer->getGuid(),
+                    base_writer->get_liveliness_kind(),
+                    base_writer->get_liveliness_lease_duration(),
                     writer_status))
         {
             EPROSIMA_LOG_ERROR(RTPS_LIVELINESS,
-                    "Could not remove writer " << W->getGuid() << " from liveliness manager");
+                    "Could not remove writer " << base_writer->getGuid() << " from liveliness manager");
         }
 
         min_manual_by_participant_ms_ = std::numeric_limits<double>::max();
@@ -794,34 +797,34 @@ bool WLP::remove_local_writer(
         manual_liveliness_assertion_->update_interval_millisec(min_manual_by_participant_ms_);
         return true;
     }
-    else if (W->get_liveliness_kind() == dds::MANUAL_BY_TOPIC_LIVELINESS_QOS)
+    else if (base_writer->get_liveliness_kind() == dds::MANUAL_BY_TOPIC_LIVELINESS_QOS)
     {
         auto it = std::find(
             manual_by_topic_writers_.begin(),
             manual_by_topic_writers_.end(),
-            W);
+            base_writer);
 
         if (it == manual_by_topic_writers_.end())
         {
-            EPROSIMA_LOG_WARNING(RTPS_LIVELINESS, "Writer " << W->getGuid() << " not found.");
+            EPROSIMA_LOG_WARNING(RTPS_LIVELINESS, "Writer " << base_writer->getGuid() << " not found.");
             return false;
         }
 
         manual_by_topic_writers_.erase(it);
 
         if (!pub_liveliness_manager_->remove_writer(
-                    W->getGuid(),
-                    W->get_liveliness_kind(),
-                    W->get_liveliness_lease_duration(),
+                    base_writer->getGuid(),
+                    base_writer->get_liveliness_kind(),
+                    base_writer->get_liveliness_lease_duration(),
                     writer_status))
         {
             EPROSIMA_LOG_ERROR(RTPS_LIVELINESS,
-                    "Could not remove writer " << W->getGuid() << " from liveliness manager");
+                    "Could not remove writer " << base_writer->getGuid() << " from liveliness manager");
         }
         return true;
     }
 
-    EPROSIMA_LOG_WARNING(RTPS_LIVELINESS, "Writer " << W->getGuid() << " not found.");
+    EPROSIMA_LOG_WARNING(RTPS_LIVELINESS, "Writer " << base_writer->getGuid() << " not found.");
     return false;
 }
 
@@ -1015,7 +1018,7 @@ void WLP::pub_liveliness_changed(
 
     if (kind == dds::AUTOMATIC_LIVELINESS_QOS)
     {
-        for (RTPSWriter* w: automatic_writers_)
+        for (BaseWriter* w: automatic_writers_)
         {
             if (w->getGuid() == writer)
             {
@@ -1035,7 +1038,7 @@ void WLP::pub_liveliness_changed(
     }
     else if (kind == dds::MANUAL_BY_PARTICIPANT_LIVELINESS_QOS)
     {
-        for (RTPSWriter* w: manual_by_participant_writers_)
+        for (BaseWriter* w: manual_by_participant_writers_)
         {
             if (w->getGuid() == writer)
             {
@@ -1055,7 +1058,7 @@ void WLP::pub_liveliness_changed(
     }
     else if (kind == dds::MANUAL_BY_TOPIC_LIVELINESS_QOS)
     {
-        for (RTPSWriter* w: manual_by_topic_writers_)
+        for (BaseWriter* w: manual_by_topic_writers_)
         {
             if (w->getGuid() == writer)
             {
