+++ src/cpp/rtps/writer/RTPSWriter.cpp
@@ -94,76 +94,6 @@ uint32_t RTPSWriter::getTypeMaxSerialized()
     return history_->getTypeMaxSerialized();
 }
 
-constexpr uint32_t info_dst_message_length = 16;
-constexpr uint32_t info_ts_message_length = 12;
-constexpr uint32_t data_frag_submessage_header_length = 36;
-constexpr uint32_t heartbeat_message_length = 32;
-
-uint32_t RTPSWriter::getMaxDataSize()
-{
-    uint32_t flow_max = flow_controller_->get_max_payload();
-    uint32_t part_max = mp_RTPSParticipant->getMaxMessageSize();
-    uint32_t max_size = flow_max > part_max ? part_max : flow_max;
-    if (max_output_message_size_ < max_size)
-    {
-        max_size = max_output_message_size_;
-    }
-
-    max_size =  calculateMaxDataSize(max_size);
-    return max_size &= ~3;
-}
-
-uint32_t RTPSWriter::calculateMaxDataSize(
-        uint32_t length)
-{
-    uint32_t maxDataSize = mp_RTPSParticipant->calculateMaxDataSize(length);
-
-    maxDataSize -= info_dst_message_length +
-            info_ts_message_length +
-            data_frag_submessage_header_length +
-            heartbeat_message_length;
-
-    //TODO(Ricardo) inlineqos in future.
-
-#if HAVE_SECURITY
-    if (getAttributes().security_attributes().is_submessage_protected)
-    {
-        maxDataSize -= mp_RTPSParticipant->security_manager().calculate_extra_size_for_rtps_submessage(m_guid);
-    }
-
-    if (getAttributes().security_attributes().is_payload_protected)
-    {
-        maxDataSize -= mp_RTPSParticipant->security_manager().calculate_extra_size_for_encoded_payload(m_guid);
-    }
-#endif // if HAVE_SECURITY
-
-#ifdef FASTDDS_STATISTICS
-    maxDataSize -= eprosima::fastdds::statistics::rtps::statistics_submessage_length;
-#endif // FASTDDS_STATISTICS
-
-    return maxDataSize;
-}
-
-const dds::LivelinessQosPolicyKind& RTPSWriter::get_liveliness_kind() const
-{
-    return liveliness_kind_;
-}
-
-const Duration_t& RTPSWriter::get_liveliness_lease_duration() const
-{
-    return liveliness_lease_duration_;
-}
-
-const Duration_t& RTPSWriter::get_liveliness_announcement_period() const
-{
-    return liveliness_announcement_period_;
-}
-
-bool RTPSWriter::is_datasharing_compatible() const
-{
-    return (m_att.data_sharing_configuration().kind() != dds::OFF);
-}
-
 }  // namespace rtps
 }  // namespace fastdds
 }  // namespace eprosima
